+++ cf_xarray/accessor.py
@@ -886,56 +886,12 @@ class _CFWrappedPlotMethods:
         )
 
 
-class CFFlagVariable:
-    def __init__(self, da):
-        self._da = da
-
-        if "flag_meanings" not in da.attrs:
-            raise ValueError(
-                "The flag_meanings attribute is absent. This is not a flag variable."
-            )
-        if "flag_values" not in da.attrs:
-            raise ValueError(
-                "The flag_values attribute is absent. This is not a flag variable."
-            )
-        flag_meanings = da.attrs["flag_meanings"].split(" ")
-        flag_values = da.attrs["flag_values"]
-        # TODO: assert flag_values is iterable
-        assert len(flag_values) == len(flag_meanings)
-        self.flag_dict = dict(zip(flag_meanings, flag_values))
-
-    def _assert_valid_other_comparison(self, other):
-        if other not in self.flag_dict:
-            raise ValueError(
-                f"Did not find flag value meaning [{other}] in known flag meanings: [{self.flag_dict.keys()!r}]"
-            )
-
-    def __eq__(self, other):
-        self._assert_valid_other_comparison(other)
-        return self._da == self.flag_dict[other]
-
-    def __lt__(self, other):
-        self._assert_valid_other_comparison(other)
-        return self._da < self.flag_dict[other]
-
-    def __le__(self, other):
-        self._assert_valid_other_comparison(other)
-        return self._da <= self.flag_dict[other]
-
-    def __gt__(self, other):
-        self._assert_valid_other_comparison(other)
-        return self._da > self.flag_dict[other]
-
-    def __ge__(self, other):
-        self._assert_valid_other_comparison(other)
-        return self._da >= self.flag_dict[other]
-
-    def isin(self, test_elements):
-        mapped_test_elements = [self.flag_dict[elem] for elem in test_elements]
-        return self._da.isin(mapped_test_elements)
-
-    def __repr__(self):
-        return f"CF Flag variable with mapping:\n\t{self.flag_dict!r}"
+def parse_flag_attrs(da):
+    flag_meanings = da.attrs["flag_meanings"].split(" ")
+    flag_values = da.attrs["flag_values"]
+    # TODO: assert flag_values is iterable
+    assert len(flag_values) == len(flag_meanings)
+    return dict(zip(flag_meanings, flag_values))
 
 
 class CFAccessor:
@@ -947,8 +903,54 @@ class CFAccessor:
         self._obj = obj
         self._all_cell_measures = None
 
-        if isinstance(self._obj, xr.DataArray):
-            self.flag = CFFlagVariable(self._obj)
+        if (
+            isinstance(self._obj, xr.DataArray)
+            and "flag_meanings" in self._obj.attrs
+            and "flag_values" in self._obj.attrs
+        ):
+            self._flag_dict = parse_flag_attrs(self._obj)
+        else:
+            self._flag_dict = None
+
+    def _assert_valid_other_comparison(self, other):
+        if not self._flag_dict:
+            raise ValueError(
+                "Comparisons are only supported for DataArrays that represent CF flag variables."
+                ".attrs must contain 'flag_values' and 'flag_meanings'"
+            )
+        if other not in self._flag_dict:
+            raise ValueError(
+                f"Did not find flag value meaning [{other}] in known flag meanings: [{self._flag_dict.keys()!r}]"
+            )
+
+    def __eq__(self, other):
+        self._assert_valid_other_comparison(other)
+        return self._obj == self._flag_dict[other]
+
+    def __lt__(self, other):
+        self._assert_valid_other_comparison(other)
+        return self._obj < self._flag_dict[other]
+
+    def __le__(self, other):
+        self._assert_valid_other_comparison(other)
+        return self._obj <= self._flag_dict[other]
+
+    def __gt__(self, other):
+        self._assert_valid_other_comparison(other)
+        return self._obj > self._flag_dict[other]
+
+    def __ge__(self, other):
+        self._assert_valid_other_comparison(other)
+        return self._obj >= self._flag_dict[other]
+
+    def isin(self, test_elements):
+        if self._flag_dict:
+            mapped_test_elements = [self._flag_dict[elem] for elem in test_elements]
+            return self._obj.isin(mapped_test_elements)
+        else:
+            raise ValueError(
+                ".cf.isin only supported DataArrays that represent CF flag variables. Please use Dataset.isin instead"
+            )
 
     def _get_all_cell_measures(self):
         """
@@ -1106,10 +1108,6 @@ class CFAccessor:
         return kwargs
 
     def __getattr__(self, attr):
-        if attr == "flag":
-            raise AttributeError(
-                "Flag variable features only apply to DataArrays, not Datasets."
-            )
         return _getattr(
             obj=self._obj,
             attr=attr,
@@ -1189,7 +1187,11 @@ class CFAccessor:
 
             return "\n".join(rows) + "\n"
 
-        text = "Coordinates:"
+        if self._flag_dict:
+            text = f"CF Flag variable with mapping:\n\t{self._flag_dict!r}\n\n"
+        else:
+            text = ""
+        text += "Coordinates:"
         text += make_text_section("CF Axes", "axes", coords, _AXIS_NAMES)
         text += make_text_section("CF Coordinates", "coordinates", coords, _COORD_NAMES)
         text += make_text_section(