+++ doc/whats-new.rst
@@ -6,6 +6,7 @@ What's New
 v0.6.1 (unreleased)
 ===================
 - Support detecting pint-backed Variables with units-based criteria. By `Deepak Cherian`_.
+- Support reshaping nD bounds arrays to (n-1)D vertex arrays. By `Deepak Cherian`_.
 
 v0.6.0 (June 29, 2021)
 ======================