+++ doc/whats-new.rst
@@ -7,7 +7,7 @@ v0.6.1 (unreleased)
 ===================
 - Support detecting pint-backed Variables with units-based criteria. By `Deepak Cherian`_.
 - Support reshaping nD bounds arrays to (n-1)D vertex arrays. By `Deepak Cherian`_.
-- Support rich comparisons and ``.isin`` for flag variables using ``DataArray.cf.flag``.
+- Support rich comparisons and ``.isin`` for flag variables with ``DataArray.cf``.
   By `Deepak Cherian`_ and `Julius Busecke`_
 
 v0.6.0 (June 29, 2021)